@@ -8,9 +8,10 @@ import '../../../common_imports.dart';
 class HomePage extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    final L10n l10n = L10n.of(context);
+    final l10n = L10n.of(context)!;
     return Scaffold(
       appBar: AppBar(
+        backgroundColor: Theme.of(context).primaryColor,
         title: Semantics(
           child: Text(l10n.homeAppbarTitle),
           label: "This is the HomePage.",
