@@ -5,7 +5,7 @@ import '../../../../common_imports.dart';
 class CurrentLanguage extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    final l10n = L10n.of(context);
+    final l10n = L10n.of(context)!;
     return Column(
       children: [
         Text(
